Introduce mutex around remote-calling code.
This is in response to an occasional SSL error that was occurring when two
remote calls - `cmd` and `files` - are made in very quick succession, when
the ASM webapp page is (re)loaded, and possibly at other times.

Whichever of the calls was made second was erroring with "certificate verify
failed" errors (despite having `SSL_VERIFY_MODE=0`). My theory is that the
second call was either trying to use, or trying to re-create, an SSL connection
that was still being set up for the first call.

The solution is to only allow one remote call at once by synchronizing on a
mutex.

This shouldn't impact multiple concurrent users of ASM since the mutex is in
the context of each API request - multiple HTTP requests to the server will use
their own `DaemonClient::Connection`s and hence `Executor`s.